refactor task file names, separated propose admin into helper files, tested locally propose-admin
