+++ hardhat.config.js
@@ -31,10 +31,10 @@ task('forta:share-type', 'Checks if a list of shares is active or inactive')
         getShareTypes({ shareIds: taskArgs.ids.split(',') });
     });
 require('./tasks/deploy');
-require('./tasks/deployAndPrepareUpgrade');
-require('./tasks/prepareUpgrade');
-require('./tasks/verifyDeployed');
-require('./tasks/proposeAdmin');
+require('./tasks/deploy-and-prepare-upgrade');
+require('./tasks/prepare-upgrade');
+require('./tasks/verify-deployed');
+require('./tasks/propose-admin');
 
 module.exports = {};
