+++ tasks/verify-deployed.js
